+++ sdk_entrepot_gpf/workflow/resolver/DictResolver.py
@@ -1,5 +1,6 @@
 from typing import Any, Dict
 
+from sdk_entrepot_gpf.io.Config import Config
 from sdk_entrepot_gpf.workflow.resolver.AbstractResolver import AbstractResolver
 from sdk_entrepot_gpf.workflow.resolver.Errors import ResolverError
 
@@ -43,5 +44,6 @@ class DictResolver(AbstractResolver):
         try:
             return str(self.get(self.__key_value, string_to_solve))
         except KeyError as e:
+            Config().om.error(f"Impossible de résoudre la clef '{string_to_solve}' pour le résolveur '{self.name}', clefs possibles au niveau 1 : {', '.join(self.__key_value.keys())}")
             # Sinon on lève une exception
             raise ResolverError(self.name, string_to_solve) from e